Uses useMiniAppDetection() and useAsciiLogoAnimation(asciiLogoLines). Replaced inline init effect with useInitializeParticipant. Replaced inline verifyTwitterFollow with useTwitterVerification(updateTask) and updated call sites.
